LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/charset-03-expected.txt


PASS UTF-8 imported module script
FAIL UTF-8 imported module script with wrong charset in Content-Type assert_equals: Should be decoded as UTF-8 expected "śćążź" but got "Ĺ›Ä‡Ä…ĹĽĹş"
PASS Non-UTF-8 imported module script
FAIL Non-UTF-8 imported module script with charset in Content-Type assert_not_equals: Should be decoded as UTF-8 got disallowed value "śćążź"
